Move item module into protocol
