+++ src/protocol/mod.rs
@@ -26,6 +26,7 @@ pub mod mojang;
 pub mod forge;
 pub mod nbt;
 pub mod format;
+pub mod item;
 
 use std::fmt;
 use std::default;
@@ -84,7 +85,7 @@ macro_rules! state_packets {
                 use crate::protocol::format;
                 use crate::protocol::nbt;
                 use crate::types;
-                use crate::item;
+                use crate::protocol::item;
                 use crate::shared::Position;
