+++ src/protocol/packet.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use super::format;
+use super::item;
 
 state_packets!(
     handshake Handshaking {
@@ -2101,7 +2102,7 @@ impl Serializable for Advancement {
 pub struct AdvancementDisplay {
     pub title: String,
     pub description: String,
-    pub icon: Option<crate::item::Stack>,
+    pub icon: Option<item::Stack>,
     pub frame_type: VarInt,
     pub flags: i32,
     pub background_texture: Option<String>,
@@ -2113,7 +2114,7 @@ impl Serializable for AdvancementDisplay {
     fn read_from<R: io::Read>(buf: &mut R) -> Result<Self, Error> {
         let title: String = Serializable::read_from(buf)?;
         let description: String = Serializable::read_from(buf)?;
-        let icon: Option<crate::item::Stack> = Serializable::read_from(buf)?;
+        let icon: Option<item::Stack> = Serializable::read_from(buf)?;
         let frame_type: VarInt = Serializable::read_from(buf)?;
         let flags: i32 = Serializable::read_from(buf)?;
         let background_texture: Option<String> = if flags & 1 != 0 {
@@ -2399,7 +2400,6 @@ pub struct PlayerProperty {
     pub signature: Option<String>,
 }
 
-use crate::item;
 type RecipeIngredient = LenPrefixed<VarInt, Option<item::Stack>>;
 
 #[derive(Debug)]
